requirements.txt
# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Queue
redis>=5.0.0

# LLM
anthropic>=0.26.0

# HTTP client (Instagram API calls)
httpx>=0.27.0

# Config
python-dotenv>=1.0.0

# Already in fwclean venv — listed for completeness / fresh installs
faster-whisper>=1.0.0
yt-dlp>=2024.1.0
